Refactored tests.
--HG--
branch : 0.7
rename : tests/test_application.py => tests/test_app.py
